pyproject.toml
[project]
name = "arktrace"
version = "0.1.0"
description = "Shadow fleet candidate screening pipeline for Cap Vista Solicitation 5.0"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.1",
    "polars>=1.0",
    "scikit-learn>=1.5",
    "shap>=0.46",
    "pyarrow>=17.0",
    "websockets>=12.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "h3>=3.7",
    "fastapi>=0.115",
    "uvicorn>=0.32",
    "lancedb>=0.9",
    "lance-graph>=0.5.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pipeline"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: opt-in integration tests requiring external/public data",
]

[dependency-groups]
dev = [
    "pip-licenses>=5.5.5",
    "playwright>=1.49",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.8",
    "mypy>=1.13",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
ignore = [
    "E501",   # line too long — handled by formatter
    "E741",   # ambiguous variable names (l, O, I) — too noisy in scientific code
]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
check_untyped_defs = true
warn_unused_ignores = true

# DuckDB connection.execute().pl() returns DataFrame | Series per type stubs
# but is always DataFrame in practice. Suppress in files that call .pl()
# extensively until DuckDB stubs are updated upstream.
[[tool.mypy.overrides]]
module = [
    "pipeline.src.features.ownership_graph",
    "pipeline.src.features.identity",
    "pipeline.src.analysis.label_propagation",
    "pipeline.src.api.routes.chat",
]
ignore_errors = true
